+++ src/index.ts
@@ -7,6 +7,7 @@ export { XmlCdata } from './lib/XmlCdata.js';
 export { XmlComment } from './lib/XmlComment.js';
 export { XmlDeclaration } from './lib/XmlDeclaration.js';
 export { XmlDocument } from './lib/XmlDocument.js';
+export { XmlDocumentType } from './lib/XmlDocumentType.js';
 export { XmlElement } from './lib/XmlElement.js';
 export { XmlNode } from './lib/XmlNode.js';
 export { XmlProcessingInstruction } from './lib/XmlProcessingInstruction.js';